+++ server/src/test/java/com/orientechnologies/orient/core/storage/impl/local/paginated/LocalPaginatedStorageSmallCacheBigRecordsCrashRestoreIT.java
@@ -20,14 +20,14 @@ import org.junit.Before;
 import org.junit.Test;
 
 import java.io.File;
+import java.io.RandomAccessFile;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
-import java.util.concurrent.Callable;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
+import java.util.concurrent.*;
 import java.util.concurrent.atomic.AtomicLong;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * @author Andrey Lomakin
@@ -46,29 +46,55 @@ public class LocalPaginatedStorageSmallCacheBigRecordsCrashRestoreIT {
     buildDirectory += "/localPaginatedStorageSmallCacheBigRecordsCrashRestore";
 
     buildDir = new File(buildDirectory);
+
+    buildDirectory = buildDir.getCanonicalPath();
+    buildDir = new File(buildDirectory);
+
     if (buildDir.exists())
       OFileUtils.deleteFolderIfEmpty(buildDir);
 
     buildDir.mkdir();
 
+    final File mutexFile = new File(buildDir, "mutex.ct");
+    final RandomAccessFile mutex = new RandomAccessFile(mutexFile, "rw");
+    mutex.seek(0);
+    mutex.write(0);
+
     String javaExec = System.getProperty("java.home") + "/bin/java";
+    javaExec = new File(javaExec).getCanonicalPath();
+
     System.setProperty("ORIENTDB_HOME", buildDirectory);
 
-    ProcessBuilder processBuilder = new ProcessBuilder(javaExec, "-Xmx2048m", "-classpath", System.getProperty("java.class.path"),
-        "-DORIENTDB_HOME=" + buildDirectory, RemoteDBRunner.class.getName());
+    ProcessBuilder processBuilder = new ProcessBuilder(javaExec, "-Xmx2048m", "-XX:MaxDirectMemorySize=512g", "-classpath",
+        System.getProperty("java.class.path"), "-DmutexFile=" + mutexFile.getCanonicalPath(), "-DORIENTDB_HOME=" + buildDirectory, RemoteDBRunner.class.getName());
     processBuilder.inheritIO();
 
     process = processBuilder.start();
 
-    Thread.sleep(5000);
+    System.out.println(LocalPaginatedStorageSmallCacheBigRecordsCrashRestoreIT.class.getSimpleName() + ": Wait for server start");
+    boolean started = false;
+    do {
+      Thread.sleep(5000);
+      mutex.seek(0);
+      started = mutex.read() == 1;
+    } while (!started);
+
+    mutex.close();
+    mutexFile.delete();
+    System.out.println(LocalPaginatedStorageSmallCacheBigRecordsCrashRestoreIT.class.getSimpleName() + ": Server was started");
   }
 
   @After
   public void tearDown() {
-    testDocumentTx.activateOnCurrentThread();
-    testDocumentTx.drop();
-    baseDocumentTx.activateOnCurrentThread();
-    baseDocumentTx.drop();
+    if (testDocumentTx != null) {
+      testDocumentTx.activateOnCurrentThread();
+      testDocumentTx.drop();
+    }
+
+    if (baseDocumentTx != null) {
+      baseDocumentTx.activateOnCurrentThread();
+      baseDocumentTx.drop();
+    }
 
     OFileUtils.deleteRecursively(buildDir);
     Assert.assertFalse(buildDir.exists());
@@ -100,13 +126,14 @@ public class LocalPaginatedStorageSmallCacheBigRecordsCrashRestoreIT {
       futures.add(executorService.submit(new DataPropagationTask(baseDocumentTx, testDocumentTx)));
     }
 
-    Thread.sleep(3000);
+    System.out.println("Wait for 5 minutes");
+    TimeUnit.MINUTES.sleep(5);
 
     long lastTs = System.currentTimeMillis();
     process.destroy();
     process.waitFor();
 
-    System.out.println("Process was destroyed");
+    System.out.println("OrientDB server process was destroyed");
 
     for (Future future : futures) {
       try {
@@ -157,23 +184,31 @@ public class LocalPaginatedStorageSmallCacheBigRecordsCrashRestoreIT {
       final ORecordId rid = new ORecordId(clusterId);
 
       for (OPhysicalPosition physicalPosition : physicalPositions) {
+
         rid.clusterPosition = physicalPosition.clusterPosition;
 
-        ODatabaseRecordThreadLocal.INSTANCE.set(baseDocumentTx);
+        baseDocumentTx.activateOnCurrentThread();
         ODocument baseDocument = baseDocumentTx.load(rid);
 
         testDocumentTx.activateOnCurrentThread();
         List<ODocument> testDocuments = testDocumentTx
             .query(new OSQLSynchQuery<ODocument>("select from TestClass where id  = " + baseDocument.field("id")));
+
         if (testDocuments.size() == 0) {
-          if (((Long) baseDocument.field("timestamp")) < minTs)
+          if (((Long) baseDocument.field("timestamp")) < minTs) {
             minTs = baseDocument.field("timestamp");
+          }
         } else {
           ODocument testDocument = testDocuments.get(0);
-          Assert.assertEquals(testDocument.field("id"), baseDocument.field("id"));
-          Assert.assertEquals(testDocument.field("timestamp"), baseDocument.field("timestamp"));
-          Assert.assertEquals(testDocument.field("stringValue"), baseDocument.field("stringValue"));
-          Assert.assertEquals(testDocument.field("binaryValue"), baseDocument.field("binaryValue"));
+
+          assertThat(testDocument.field("id")).as("id:: %s", testDocument.field("id")).isEqualTo(baseDocument.field("id"));
+          assertThat(testDocument.field("timestamp")).as("documents:: %s - %s", testDocument, baseDocument)
+              .isEqualTo(baseDocument.field("timestamp"));
+          assertThat(testDocument.field("stringValue")).as("id:: %s", testDocument.field("id"))
+              .isEqualTo(baseDocument.field("stringValue"));
+          assertThat(testDocument.field("binaryValue")).as("id:: %s", testDocument.field("id"))
+              .isEqualTo(baseDocument.field("binaryValue"));
+
           recordsRestored++;
         }
 
@@ -200,6 +235,11 @@ public class LocalPaginatedStorageSmallCacheBigRecordsCrashRestoreIT {
           .getResourceAsStream("/com/orientechnologies/orient/core/storage/impl/local/paginated/db-create-big-records-config.xml"));
       server.activate();
 
+      final String mutexFile = System.getProperty("mutexFile");
+      final RandomAccessFile mutex = new RandomAccessFile(mutexFile, "rw");
+      mutex.seek(0);
+      mutex.write(1);
+      mutex.close();
     }
   }
 
@@ -234,7 +274,11 @@ public class LocalPaginatedStorageSmallCacheBigRecordsCrashRestoreIT {
         }
 
       } finally {
+        baseDB.activateOnCurrentThread();
+
         baseDB.close();
+        testDB.activateOnCurrentThread();
+
         testDB.close();
       }
     }
